Pass objects by unique_ptr in accordance with create
create factory methods cannot return a value in order to support
polymorphism. So they have to return dynamically allocated data, some
sort of pointers, unique_ptr is a good choice. But for getters it is
inconvenient to return unique_ptr, and using another type of smart
pointers doesn't help to keep things consistent. There is also the
implicit restriction that the user must never copy returned reference
to his objects because of polymorphism. So an abstraction is needed to
solve that.